[IMP] tag_editable_requirements: tag by egg name
Include the egg name in tags of editable requirements, so we
can have different versions of different odoo addons in the
same repository. The only drawback is that it creates several tags
when you have several dependencies in the same repo, even if they
are on the same sha. The benefit is that it gives more flexibility
and also explicitly show why the tag is present by including the
addon name in the tag name.